+++ NEWS.md
@@ -1,3 +1,19 @@
+# rankinPlot 1.1
+
+grottaBar() function has several new options:
+
+* Printing of count (percent) format for information
+* The ability to specify a font face (plain, bold, italic) for the numbers printed in the Grotta Bar.
+* The ability to specify two colours for the numbers printed in the Grotta Bar.
+
+Examples of these options may be found in the documentation for the grottaBar function.
+
+Several other quality of life changes have been added:
+
+* The package now supports both spellings of "colour".
+* printNumbers options may now be abbreviated
+
+
 # rankinPlot 1.0
 
 Package released